Bullseye: minor fix

+++ modules/posbullseye/templates/iptables_apache_accept.sh
@@ -7,7 +7,9 @@ iptables -D INPUT -p tcp --dport 443 -j DROP
 iptables -D INPUT -p tcp --dport 80 -j DROP
 
 #delete temporary rules
-rm /etc/iptables/rules.v4
+if [[ -f /etc/iptables/rules.v4 ]]; then
+  rm /etc/iptables/rules.v4
+fi
 
 #list iptables
 iptables -L
